Property test tidy-ups
