+++ packages/pbt-properties/test/PropertyResult.test.ts
@@ -1,11 +1,15 @@
+import fc from 'fast-check';
 import * as dev from '../src';
+import * as devCore from 'pbt-core';
 import * as stable from './helpers/stableApi';
 import {
   extendableArbitrary,
-  arbitraryPropertyConfig,
   arbitraryGens,
   arbitrarySucceedingPropertyFunction,
+  arbitraryPropertyConfig,
+  arbitrarilyShuffleArray,
 } from './helpers/arbitraries';
+import { GenStub } from './helpers/stubs';
 
 test('Given a succeeding property function, the property holds', () => {
   const arb = extendableArbitrary()
@@ -39,6 +43,36 @@ test('Given a false predicate, the property does not hold', () => {
       const result = p(config);
 
       expect(result).toMatchObject({ kind: 'failure', problem: { kind: 'predicate' } });
+    }),
+  );
+});
+
+test('Given an exhausting generator, the property does not hold', () => {
+  const arb = extendableArbitrary()
+    .extend(() => arbitraryPropertyConfig())
+    .extend(
+      ({ iterations }) =>
+        arbitraryGens({ minLength: iterations })
+          .map((gs) => [...gs, GenStub.exhausted()])
+          .chain(arbitrarilyShuffleArray) as fc.Arbitrary<devCore.Gens>,
+    )
+    .extend(() => arbitrarySucceedingPropertyFunction())
+    .toArbitrary();
+
+  stable.assert(
+    stable.property(arb, ([config, gs, f]) => {
+      const p = dev.property(...gs, f);
+
+      const result = p(config);
+
+      expect(result).toEqual({
+        kind: 'failure',
+        problem: {
+          kind: 'exhaustion',
+          iterationsRequested: config.iterations,
+          iterationsCompleted: 0,
+        },
+      });
     }),
   );
 });
