+++ packages/pbt-properties/test/Validation.test.ts
